requirements.txt
coverage==4.0.2
coveralls==1.1
defusedxml==0.4.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.6
django-bootstrap3==6.2.2
django-cors-headers==1.1.0
django-debug-toolbar==1.4
django-extensions==1.5.9
django-toolbelt==0.0.1
djangorestframework==3.3.1
docopt==0.6.2
gunicorn==19.3.0
nose==1.3.7
oauthlib==1.0.3
Pillow==3.0.0
psycopg2==2.6.1
PyJWT==1.4.0
python-memcached==1.57
python-social-auth==0.2.13
python3-openid==3.0.9
requests==2.8.1
requests-oauthlib==0.5.0
six==1.10.0
sqlparse==0.1.18
static3==0.6.1

wheel==0.24.0